+++ packages/gamut-labs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [33.0.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@33.0.3...@codecademy/gamut-labs@33.0.4) (2022-07-29)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [33.0.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@33.0.2...@codecademy/gamut-labs@33.0.3) (2022-07-29)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [33.0.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@33.0.1...@codecademy/gamut-labs@33.0.2) (2022-07-27)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
 ### [33.0.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@33.0.0...@codecademy/gamut-labs@33.0.1) (2022-07-27)
 
 **Note:** Version bump only for package @codecademy/gamut-labs