+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [14.0.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@14.0.1...@codecademy/gamut-styles@14.0.2) (2022-07-29)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+### [14.0.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@14.0.0...@codecademy/gamut-styles@14.0.1) (2022-07-29)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+## [14.0.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@13.4.0...@codecademy/gamut-styles@14.0.0) (2022-07-27)
+
+
+### ⚠ BREAKING CHANGES
+
+* **Text:** highlighted text
+
+### Features
+
+* **Text:** highlighted text ([4b1d299](https://github.com/Codecademy/gamut/commit/4b1d2995267e90fe21cb7934598f140963fd4780))
+
+
+
 ## [13.4.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@13.3.0...@codecademy/gamut-styles@13.4.0) (2022-07-19)